Finalized /factcheck command
Implemented the /factcheck "query" command in the Discord bot. This command allows users to input a string query, which the bot processes to retrieve and display an embed containing a list of false claims related to the query. The data is fetched using the FactChecker API, filtered for claims with a 'False' rating. The response is presented in an embed format within Discord, offering users an easy and interactive way to fact-check statements or claims.